Add 'role_arn' to 'aws_eks_cluster'
The 'role_arn' has been added to the 'aws_eks_cluster' in the 'eks.tf' file. The 'role_arn' attribute provides an Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that provides permissions for the Kubernetes control plane to make calls to AWS API operations on your behalf.