Remove VPC data block from 'eks.tf'
The data block for the existing VPC has been removed from the 'eks.tf' file. Now, instead of pulling the VPC ID from the data block, the ID is directly provided within the 'eks_sg' resource block. This aims to simplify the configurations and eliminate dependencies on specific data blocks.